api_yamdb/requirements.txt
asgiref==3.2.10
atomicwrites==1.4.1
attrs==19.3.0
certifi==2020.4.5.1
chardet==3.0.4
colorama==0.4.5
Django==3.0.5
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
gunicorn==20.0.4
idna==2.9
importlib-metadata==1.6.0
more-itertools==8.2.0
packaging==20.3
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.8.1
PyJWT==2.1.0
pyparsing==2.4.7
pytest==5.4.1
pytest-django==3.9.0
pytest-pythonpath==0.7.4
pytz==2020.1
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.9
wcwidth==0.1.9
zipp==3.1.0